Update docker compose to not bind to native port by default
This makes some changes to make docker not specifically bind to the
native/default `3306` port for mysql.  This allows running multiple
database container images without conflicting.

the .env.sample file is updated with a DATABASE_URL configuration
that will read the correct path / port out of the running docker image,
so this can happen seamlessly.